Initial Commit
Base implementation of the WhatchaSay Dalamud Plugin.

Used the Discord Bridge as an example and will have to further review the code as a lot of it is temporarily copied.

The plugin can translate any non-system message technically however it works best on chat.

Users will have to provide their own API Key if they use the DeepL option. I will also be adding the ability to choose other providers for LibreTranslate as well as a custom domain.